backend/backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
fonttools==4.56.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
Naked==0.1.32
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.3.2
pillow==11.1.0
pycparser==2.22
pycryptodome==3.21.0
pygame==2.5.2
PyJWT==2.9.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
